+++ CHANGELOG.md
@@ -6,11 +6,12 @@ This changelog only contains the changes that are unreleased. For changes for in
 ## 3.4.13.7
 
 ### New Features
+- When exporting to Modrinth format, check all mods on Modrinth
+- When exporting to CurseForge format, check all mods on CurseForge
 
 ### Fixes
 - Remove caching of download urls from CurseForge as per ToS, also remove some others from json
 - Don't export CurseForge mods urls in Modrinth pack exports due to CurseForge ToS, use overides
-- When exporting to Modrinth format, check all mods on Modrinth
 - Change right click menu on edit mods for mods with info from both mod platforms
 
 ### Misc